ZK-1337: change folder structure & remove console logs
