+++ ts/shielder-sdk/src/actions/index.ts
@@ -1,6 +1,7 @@
 import { NewAccountAction, NewAccountCalldata } from "./newAccount";
 import { DepositAction, DepositCalldata } from "./deposit";
 import { WithdrawAction, WithdrawCalldata } from "./withdraw";
+import { INonceGenerator } from "./utils";
 
 export interface Calldata {
   provingTimeMillis: number;
@@ -12,5 +13,6 @@ export {
   NewAccountAction,
   NewAccountCalldata,
   WithdrawAction,
-  WithdrawCalldata
+  WithdrawCalldata,
+  INonceGenerator
 };
