+++ ts/shielder-sdk/src/index.ts
@@ -2,7 +2,7 @@ export {
   createShielderClient,
   SendShielderTransaction,
   OutdatedSdkError
-} from "@/shielder/client";
-export { type AccountState, type ShielderTransaction } from "@/shielder/state";
-export { storageSchema, InjectedStorageInterface } from "@/shielder/state";
+} from "@/client";
+export { type AccountState, type ShielderTransaction } from "@/state";
+export { storageSchema, InjectedStorageInterface } from "@/state";
 export { shieldActionGasLimit } from "@/constants";
