+++ ts/shielder-sdk/src/state/index.ts
@@ -1,0 +1,20 @@
+import { StateManager } from "./manager";
+import { StateSynchronizer, UnexpectedVersionInEvent } from "./sync";
+import { StateEventsFilter } from "./events";
+import storageSchema, {
+  InjectedStorageInterface,
+  createStorage
+} from "./storageSchema";
+import { ShielderTransaction, AccountState } from "./types";
+
+export {
+  storageSchema,
+  createStorage,
+  AccountState,
+  InjectedStorageInterface,
+  ShielderTransaction,
+  StateManager,
+  StateSynchronizer,
+  StateEventsFilter,
+  UnexpectedVersionInEvent
+};
